Refactor page folder structure
