+++ src/App.tsx
@@ -9,20 +9,20 @@ import { OrderbookProvider } from './contexts/orderbook'
 import NavBar from './components/NavBar'
 import SideBar from './components/SideBar'
 
-import Factory from './pages/Factory'
+import Factory from './pages/Protocol/Factory'
 import ProtocolHome from './pages/Protocol'
 import Account from './pages/Account'
-import AccountVault from './pages/AccountVaults'
-import ConnectWallet from './pages/ConnectWallet'
-import Vault from './pages/VaultDetail'
-import Oracle from './pages/Oracle'
-import Operators from './pages/Operators'
+import AccountVault from './pages/Account/AccountVaults'
+import ConnectWallet from './pages/Account/ConnectWallet'
+import Vault from './pages/Account/VaultDetail'
+import Oracle from './pages/Protocol/Oracle'
+import Operators from './pages/Account/Operators'
 import HomePage from './pages/HomePage'
 import Settings from './pages/Settings'
 import Trade from './pages/Trade'
 
-import Orderbook from './pages/OrderBookTrade'
-import Swap from './pages/Swap'
+import Orderbook from './pages/Trade/OrderBookTrade'
+import Swap from './pages/Trade/Swap'
 import { useConnection } from './hooks/useConnection'
 
 import { getPreference } from './utils/storage'
