+++ src/pages/Protocol/Oracle/index.tsx
@@ -1,19 +1,18 @@
 import React, { useState, useEffect, useMemo } from 'react'
 import { Header, DataView, DropDown, useToast, Tag, Help } from '@aragon/ui'
 import BigNumber from 'bignumber.js'
-import LabelText from '../../components/LabelText'
-import CustomIdentityBadge from '../../components/CustomIdentityBadge'
-import { useConnectedWallet } from '../../contexts/wallet'
-import { useAsyncMemo } from '../../hooks/useAsyncMemo'
-import { expiryToDate, toTokenAmount } from '../../utils/math'
-import { getOracleAssetsAndPricers } from '../../utils/graph'
-import SectionTitle from '../../components/SectionHeader'
+import LabelText from '../../../components/LabelText'
+import CustomIdentityBadge from '../../../components/CustomIdentityBadge'
+import { useConnectedWallet } from '../../../contexts/wallet'
+import { useAsyncMemo } from '../../../hooks/useAsyncMemo'
+import { expiryToDate, toTokenAmount } from '../../../utils/math'
+import { getOracleAssetsAndPricers } from '../../../utils/graph'
+import SectionTitle from '../../../components/SectionHeader'
 
-import { SubgraphPriceEntry } from '../../types'
-// import { CTokenPricer, USDCPricer, CLPricer } from '../../utils/contracts/pricers'
+import { SubgraphPriceEntry } from '../../../types'
 import { pricerMap } from './config'
-import { ZERO_ADDR } from '../../constants/addresses'
-import { PRICE_SUBMISSION } from '../../constants/dataviewContents'
+import { ZERO_ADDR } from '../../../constants/addresses'
+import { PRICE_SUBMISSION } from '../../../constants/dataviewContents'
 
 export default function Oracle() {
   const { networkId } = useConnectedWallet()
@@ -21,9 +20,6 @@ export default function Oracle() {
   const [isLoadingHistory, setIsLoadingHistory] = useState(true)
   const [selectedAssetIndex, setSelectedAssetIndex] = useState(-1)
   const [assetHistory, setAssetHistory] = useState<SubgraphPriceEntry[]>([])
-  // const [isSearchingID, setIsSearching] = useState(false)
-
-  // const [expiryIdxToSubmit, setExpiryIdxToSubmit] = useState(-1)
 
   const allOracleAssets = useAsyncMemo(
     async () => {
@@ -36,21 +32,6 @@ export default function Oracle() {
     [],
   )
 
-  // const allOTokens = useAsyncMemo(async () => {
-  //   const oTokens = await getOTokens(networkId, toast)
-  //   return oTokens === null ? [] : oTokens
-  // }, [], [])
-
-  // const unsetExpiries = useMemo(() => {
-  //   const alreadySet = assetHistory.map(entry => Number(entry.expiry))
-  //   const unique = new Set(allOTokens
-  //     .map(o => Number(o.expiryTimestamp))
-  //     .filter(expiry => expiry < Date.now() / 1000)
-  //     .filter(expiry => !alreadySet.includes(expiry)))
-  //   return Array.from(unique)
-
-  // }, [assetHistory, allOTokens])
-
   const haveValidSelection = useMemo(() => allOracleAssets.length > 0 && selectedAssetIndex !== -1, [
     allOracleAssets,
     selectedAssetIndex,
@@ -61,29 +42,6 @@ export default function Oracle() {
     if (!haveValidSelection) return
     setAssetHistory(allOracleAssets[selectedAssetIndex].prices)
   }, [selectedAssetIndex, allOracleAssets, haveValidSelection])
-
-  // const setPrice = useCallback(
-  //   () => {
-  //     const selectedAsset = allOracleAssets[selectedAssetIndex].asset
-  //     const pricer = allOracleAssets[selectedAssetIndex].pricer.id
-  //     if(web3 === null) {
-  //       toast('Please connect wallet first')
-  //       return
-  //     }
-  //     if (pricerMap[selectedAsset.symbol] === PricerTypes.CTokenPricer) {
-  //       const contract = new CTokenPricer(web3, pricer , networkId, user)
-  //       return contract.setPrice(unsetExpiries[expiryIdxToSubmit].toString())
-  //     } else if (pricerMap[selectedAsset.symbol] === PricerTypes.USDCPricer) {
-  //       const contract = new USDCPricer(web3, pricer, networkId, user)
-  //       return contract.setPrice(unsetExpiries[expiryIdxToSubmit].toString())
-  //     } else {
-  //       const contract = new CLPricer(web3, pricer, networkId, user)
-  //       toast('Fetching Data from Chainlink Oracle, this could take a while...')
-  //       return contract.setPrice(unsetExpiries[expiryIdxToSubmit].toString(), toast, setIsSearching)
-  //     }
-  //   },
-  //   [allOracleAssets, selectedAssetIndex, expiryIdxToSubmit, networkId, user, web3, unsetExpiries, toast],
-  // )
 
   return (
     <>
@@ -131,29 +89,6 @@ export default function Oracle() {
           </div>
         </div>
       </div>
-      {/* <SectionTitle title="Submit Price" />
-
-      <div style={{ display: 'flex', alignItems: 'center', paddingBottom: '3%' }}>
-        <div style={{ width: '30%' }}>
-          <DropDown 
-            placeholder={unsetExpiries.length > 0 ? 'Choose Expiry' : 'No unset expiry'}
-            disabled={unsetExpiries.length===0}
-            items={unsetExpiries ? unsetExpiries.map(expiry => expiryToDate(expiry)) : []}
-            selected={expiryIdxToSubmit}
-            onChange={setExpiryIdxToSubmit}
-          />
-        </div>
-
-        <div style={{ width: '30%' }}>
-          <LabelText label=' ' /> 
-          <Button 
-            disabled={!haveValidSelection || expiryIdxToSubmit === -1}
-            label={isSearchingID ? 'Checking Oracle...' : 'Set Price'}
-            onClick={setPrice}
-            icon={isSearchingID ? <LoadingRing /> : null}
-          />
-        </div>
-      </div> */}
       <SectionTitle title="Price Submissions" />
       <DataView
         status={isLoadingHistory ? 'loading' : 'default'}
