+++ src/pages/Trade/OrderBookTrade/MarketTicket.tsx
@@ -1,24 +1,24 @@
 import React, { useMemo, useState, useCallback, useEffect } from 'react'
 import BigNumber from 'bignumber.js'
 import { Button, TextInput, IconArrowRight, IconUnlock, Timer } from '@aragon/ui'
-import { SubgraphOToken, SignedOrder, OTokenBalance } from '../../types'
+import { SubgraphOToken, SignedOrder, OTokenBalance } from '../../../types'
 
-import { calculateOrderInput, calculateOrderOutput } from '../../utils/0x-utils'
-import { toTokenAmount, fromTokenAmount } from '../../utils/math'
+import { calculateOrderInput, calculateOrderOutput } from '../../../utils/0x-utils'
+import { toTokenAmount, fromTokenAmount } from '../../../utils/math'
 
-import { TradeAction, Errors } from '../../constants'
-import { useConnectedWallet } from '../../contexts/wallet'
-import { getUSDC, addresses } from '../../constants/addresses'
+import { TradeAction, Errors } from '../../../constants'
+import { useConnectedWallet } from '../../../contexts/wallet'
+import { getUSDC, addresses } from '../../../constants/addresses'
 
-import oETHIcon from '../../imgs/oETH.svg'
-import USDCIcon from '../../imgs/USDC.png'
-import { useOrderbook } from '../../contexts/orderbook'
-import { use0xExchange } from '../../hooks/use0xExchange'
-import { useUserAllowance } from '../../hooks/useAllowance'
-import { simplifyOTokenSymbol } from '../../utils/others'
-import WarningText from '../../components/Warning'
-import LabelText from '../../components/LabelText'
-import TokenBalanceEntry from '../../components/TokenBalanceEntry'
+import oETHIcon from '../../../imgs/oETH.svg'
+import USDCIcon from '../../../imgs/USDC.png'
+import { useOrderbook } from '../../../contexts/orderbook'
+import { use0xExchange } from '../../../hooks/use0xExchange'
+import { useUserAllowance } from '../../../hooks/useAllowance'
+import { simplifyOTokenSymbol } from '../../../utils/others'
+import WarningText from '../../../components/Warning'
+import LabelText from '../../../components/LabelText'
+import TokenBalanceEntry from '../../../components/TokenBalanceEntry'
 
 enum Updates {
   Input,
