+++ src/pages/Trade/OrderBookTrade/Orderbook.tsx
@@ -1,15 +1,15 @@
 import React, { useMemo, useCallback, useState } from 'react'
 
 import { DataView, Timer, Checkbox } from '@aragon/ui'
-import { SubgraphOToken, OrderWithMetaData } from '../../types'
-import { useOrderbook } from '../../contexts/orderbook'
-import { getAskPrice, getBidPrice, getRemainingAmounts } from '../../utils/0x-utils'
+import { SubgraphOToken, OrderWithMetaData } from '../../../types'
+import { useOrderbook } from '../../../contexts/orderbook'
+import { getAskPrice, getBidPrice, getRemainingAmounts } from '../../../utils/0x-utils'
 import { green, red } from './StyleDiv'
-import { toTokenAmount } from '../../utils/math'
-import { generateNoOrderContent, NO_TOKEN_SELECTED } from '../../constants/dataviewContents'
-import { TradeAction } from '../../constants'
-import { simplifyOTokenSymbol } from '../../utils/others'
-import { getPreference, storePreference } from '../../utils/storage'
+import { toTokenAmount } from '../../../utils/math'
+import { generateNoOrderContent, NO_TOKEN_SELECTED } from '../../../constants/dataviewContents'
+import { TradeAction } from '../../../constants'
+import { simplifyOTokenSymbol } from '../../../utils/others'
+import { getPreference, storePreference } from '../../../utils/storage'
 
 const SHOW_BOTH_KEY = 'orderbook-show-all'
